# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/mailing_ab_tests:
    get:
      tags:
        - MailingAbTests
      summary: List MailingAbTests
      description: Get a paginated list of MailingAbTests with optional filtering and sorting
      operationId: listMailingAbTests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingAbTestList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MailingAbTests
      summary: Create MailingAbTest
      description: Create a new MailingAbTest
      operationId: createMailingAbTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMailingAbTestRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingAbTest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/mailing_audiences:
    get:
      tags:
        - MailingAudiences
      summary: List MailingAudiences
      description: Get a paginated list of MailingAudiences with optional filtering and sorting
      operationId: listMailingAudiences
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingAudienceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MailingAudiences
      summary: Create MailingAudience
      description: Create a new MailingAudience
      operationId: createMailingAudience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMailingAudienceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingAudience'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    MailingAbTest:
      type: object
      required:
        - id
        - campaign_id
        - winner_selection
        - completed
        - sampling_seed
        - metadata
      properties:
        id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        winner_selection:
          type: string
        promote_at:
          nullable: true
          type: string
          format: date-time
        completed:
          type: boolean
        winner_mailing_id:
          nullable: true
          type: string
          format: uuid
        sampling_seed:
          type: string
          maxLength: 64
        metadata:
          type: object
...
    CreateMailingAbTestRequest:
      type: object
      required:
        - campaign_id
        - winner_selection
        - completed
        - sampling_seed
        - metadata
      properties:
        campaign_id:
          type: string
          format: uuid
        winner_selection:
          type: string
        promote_at:
          nullable: true
          type: string
          format: date-time
        completed:
          type: boolean
        winner_mailing_id:
          nullable: true
          type: string
          format: uuid
        sampling_seed:
          type: string
          maxLength: 64
        metadata:
          type: object
...
    MailingAbTestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MailingAbTest'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    MailingAudience:
      type: object
      required:
        - id
        - name
        - active
        - is_public
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        active:
          type: boolean
        is_public:
          type: boolean
        metadata:
          type: object
...
    CreateMailingAudienceRequest:
      type: object
      required:
        - name
        - active
        - is_public
        - metadata
      properties:
        name:
          type: string
          maxLength: 120
        active:
          type: boolean
        is_public:
          type: boolean
        metadata:
          type: object
...
    MailingAudienceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MailingAudience'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []